requirements.txt
asgiref==3.5.2
Django==4.0.3
djangorestframework==3.13.1
numpy==1.23.4
pandas==1.5.1
Pillow==9.0.1
python-dateutil==2.8.2
pytz==2022.6
six==1.16.0
sqlparse==0.4.3
tzdata==2022.6
